api/requirements-api.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
fastapi==0.63.0
fastapi-utils==0.2.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
importlib-metadata==4.0.1
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.2.0
pydantic==1.8.1
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.15
starlette==0.13.6
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
zipp==3.4.1
